Fix test names, share resolve_dataset via table_identifier module
- Rename test_array_ipc_roundtrip_* to test_query_key_ipc_roundtrip_*
- Move resolve_dataset to table_identifier.rs, use from both
  ann_proto.rs and filtered_read_proto.rs
